Expose the contents of metadata.yaml files to the Jinja templates and display this metadata in the generated documentation
The maintainer, dependencies, supported platforms and tips on how to use the
the framework with CMake and QMake is displayed in the menu. There is also a
link to #inqlude, in order to emphasize that KDE frameworks are integrated and
usable in the Qt ecosystem.

The dependencies are not very precise (they are based solely on the Tier in
which the framework is), but listing the complete list of dependencies should
be possible once more advanced tools are available.